Modified Container and LiquidContainer classes. Added Fridge and gas containers classes.

+++ APBD_containers_zaj3/Classes/Container.cs
@@ -2,17 +2,17 @@ using APBD_containers_zaj3.Exceptions;
 
 namespace APBD_containers_zaj3.Classes;
 
-public abstract class Container (string containerType, int cargoWeight, int containerHeight, 
-    int containerWeight, int containerDepth, int maxCargoWeight)
+public abstract class Container (string containerType, double cargoWeight, int containerHeight, 
+    double containerWeight, int containerDepth, double maxCargoWeight)
 {
     private static int _numeralId = 0;
 
     public string SerialNumber { get; } = $"KON-{containerType}-{_numeralId++}";
-    public int CargoWeight { get; protected set; } = cargoWeight;
+    public double CargoWeight { get; protected set; } = cargoWeight;
     public int ContainerHeight { get; protected set; } = containerHeight;
-    public int ContainerWeight { get; protected set; } = containerWeight;
+    public double ContainerWeight { get; protected set; } = containerWeight;
     public int ContainerDepth { get; protected set; } = containerDepth;
-    public int MaxCargoWeight { get; protected set; } = maxCargoWeight;
+    public double MaxCargoWeight { get; protected set; } = maxCargoWeight;
 
     public virtual void EmptyContainer()
     {
@@ -20,7 +20,7 @@ public abstract class Container (string containerType, int cargoWeight, int cont
         Console.WriteLine($"Emptied the container {SerialNumber}");
     }
 
-    public virtual void AddCargo(int mass)
+    public virtual void AddCargo(double mass)
     {
         if (mass + CargoWeight > MaxCargoWeight)
             throw new OverfillException("WARNING! Overfill is inevitable!");
